pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "vortex-cli"
dynamic = ["version"]
description = "CLI для управление Vortex системой"
requires-python = ">=3.9"
dependencies = [
    "psycopg>=3.1.0",
    "python-dotenv>=1.0.0",
    "rich>=10.11.0",
    "prompt_toolkit>=3.0.36",
    "tomli>=2.0.1; python_version < '3.11'"
]

[project.scripts]
vortex = "vortex.core.cli:main"

[tool.setuptools]
packages = ["vortex", "vortex.config", "vortex.core", "vortex.database", "vortex.ui"]

[tool.setuptools.dynamic]
version = {attr = "vortex.config.manager.VERSION"}
